% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multimapping.R
\name{alignFasta}
\alias{alignFasta}
\title{calls STAR on fasta}
\usage{
alignFasta(
  fasta_file,
  prefix = sub("fasta", "", fasta_file),
  cleanup = c("Log.out", "Log.progress.out", "SJ.out.tab"),
  star_path = "/slipstream/galaxy/production/galaxy-dist/tools/star/STAR",
  index_path = "/slipstream/galaxy/data/hg38/star_index"
)
}
\arguments{
\item{fasta_file}{fasta or fastq file to align}

\item{prefix}{prefix for STAR output}

\item{cleanup}{suffixes of files to remove.  by default only sam is kept.}

\item{star_path}{path to STAR executable}

\item{index_path}{path to genome indexes for STAR}
}
\description{
calls STAR on fasta
}
