% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{annotate_gr}
\alias{annotate_gr}
\title{my_annotate}
\usage{
annotate_gr(gr, tss_gr, max_dist = 5000)
}
\arguments{
\item{gr_size}{}
}
\value{

}
\description{
my_annotate
}
\examples{
tx_gr = rtracklayer::import.gff("~/gencode.v36.annotation.gtf",
feature.type = "transcript", format = "gtf")
tss_gr = GenomicRanges::promoters(tx_gr, 1, 0)

qgr = CTCF_in_10a_overlaps_gr
    
annotate_gr(qgr, tss_gr, max_dist = 1e4)
}
