% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{annotate_gr_clusters}
\alias{annotate_gr_clusters}
\title{annotate_gr_clusters}
\usage{
annotate_gr_clusters(qgr, clust_assign, tss_gr, max_dist = 5000)
}
\arguments{
\item{qgr}{}

\item{clust_assign}{A data.frame derived from ssvSignalClustering output (clust_assign = unique(clust_dt[, .(id, cluster_id)]))}
}
\value{

}
\description{
annotate_gr_clusters
}
\examples{
clust_dt = ssvSignalClustering(CTCF_in_10a_profiles_dt)
clust_dt.assign = unique(clust_dt[, .(id, cluster_id)])
qgr = CTCF_in_10a_overlaps_gr

tx_gr = rtracklayer::import.gff("~/gencode.v36.annotation.gtf",
                                feature.type = "transcript", format = "gtf")
tss_gr = GenomicRanges::promoters(tx_gr, 1, 0)

annotate_gr(qgr, tss_gr, max_dist = 1e4)
gls = annotate_gr_clusters(qgr, clust_dt.assign, tss_gr, max_dist = 1e4)
}
