% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_bfcif.R
\name{bfcif}
\alias{bfcif}
\title{bfcif}
\usage{
bfcif(bfc, rname, FUN, force_overwrite = FALSE,
  return_path_only = FALSE, verbose = TRUE)
}
\arguments{
\item{return_path_only}{boolean, if TRUE, FUN is never run and cache_path
is returned.  if file.exist(cache_path) == TRUE, data has already been
cached.}

\item{verbose}{boolean, if TRUE status messages are displayed}
}
\description{
bfcif
}
