% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_file_helpers.R
\name{cache_FUN}
\alias{cache_FUN}
\title{Title}
\usage{
cache_FUN(x, input_name, output_tag, FUN, input_obj = input_name,
  saveFUN = saveRDS, loadFUN = readRDS)
}
\arguments{
\item{x}{a bioc cache}

\item{input_name}{path to file FUN works on}

\item{output_tag}{character.  Used to create cache result identifier (rname).
New rname
will be either 1) the result of adding this tag to rname of input_name
if input_name is cached or 2) the result of adding this tag to basename of
input_name if uncached.  If this new rname exists in cache, cached result
will be used and FUN will not be rerun.}

\item{FUN}{function that accepts one file path as it's only argument.
the output of FUN(input_name) will be returned, either from running or cached
if available.}
}
\value{
object returned by FUN(input_obj)
}
\description{
Title
}
