% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_file_helpers.R
\name{cache_gz}
\alias{cache_gz}
\title{cache_gz}
\usage{
cache_gz(x, path)
}
\arguments{
\item{x}{a BiocFileCache cache}

\item{path}{url or path to gzipped resource}

\item{rname}{rname for gzipped resource}
}
\value{
path to gunzipped resource in cache
}
\description{
Automatically handles cacheing of (potentially) gzipped resources.  Resources
can be local or remote (will be downloaded and cached).  If resource is
gzipped (only if ends in .gz) a gunzipped version will be added to cache.
}
\details{
for a given path
1. checks if it's a url and downloads, and updates to local path.
2. if path ends in .gz, gunzips and updates path
}
