% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corrDiff.R
\name{corr_fragLens}
\alias{corr_fragLens}
\title{compares correlation of specified fragLens to un-extended read pileup}
\usage{
corr_fragLens(bam_file, qgr, fragLens = 200)
}
\arguments{
\item{fragLens}{}
}
\value{
data.table merge with qgr with read and fragment correlation
}
\description{
compares correlation of specified fragLens to un-extended read pileup
}
