% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corrDiff.R
\name{corr_fragLens.single}
\alias{corr_fragLens.single}
\title{returns correlation scores for all regions in qgr as specified fragLens}
\usage{
corr_fragLens.single(bam_file, qgr, fragLens = 200)
}
\arguments{
\item{fragLens}{}
}
\value{
data.table with id and correlation
}
\description{
returns correlation scores for all regions in qgr as specified fragLens
}
