% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_RNAseq_counts.R
\name{counts_from_bams}
\alias{counts_from_bams}
\title{Title}
\usage{
counts_from_bams(bam_paths, p = 8, gtf_path = HG38_v28_GTF_URL,
  counts_tag = "counts", cache_path = "~/.cache")
}
\description{
Title
}
