% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_fix_xlsx_dates.R
\name{fix_xlsx_dates}
\alias{fix_xlsx_dates}
\title{fix_xlsx_dates}
\usage{
fix_xlsx_dates(input, targets = c("Mar", "Sep"), fixes = c("March", "Sept"))
}
\arguments{
\item{input}{character or factor to repair}

\item{targets}{date strings to search for}

\item{fixes}{replacement character when targets found.  must parallel targets.}
}
\value{
same class as input with dates converted back to gene names
}
\description{
repair any gene names that excel has converted to dates
}
