% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scRNAseq.R
\name{gene_name_hs2mm}
\alias{gene_name_hs2mm}
\title{gene_name_gs2mm}
\usage{
gene_name_hs2mm(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
converts a human gene name to mouse gene name ie. YFG -> Yfg
this is an informal conversion that mostly works but not ideal.
}
