% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_scatter_with_side_density.R
\name{get_gg_xrange}
\alias{get_gg_xrange}
\title{returns x limits of a ggplot}
\usage{
get_gg_xrange(p)
}
\arguments{
\item{p}{a ggplot}
}
\value{
x limits
}
\description{
returns x limits of a ggplot
}
