% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_scatter_with_side_density.R
\name{get_gg_yrange}
\alias{get_gg_yrange}
\title{returns y limits of a ggplot}
\usage{
get_gg_yrange(p)
}
\arguments{
\item{p}{a ggplot}
}
\value{
y limits
}
\description{
returns y limits of a ggplot
}
