% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_hclust.R
\name{get_hclust_gls}
\alias{get_hclust_gls}
\title{get_hclust_gls}
\usage{
get_hclust_gls(hclust_res, nclust)
}
\arguments{
\item{hclust_res}{result from hclust().}

\item{nclust}{number of clusters to cutree with.}
}
\value{

}
\description{
All of this is required to get cluster gene lists in the "as plotted" order with cluster 1 at the top.  
Cluster numbers when cutting trees are normally arbitrary. 
Also gene order when cutting tree would not normally be "as plotted"
}
