% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scRNAseq.R
\name{get_meta_dt}
\alias{get_meta_dt}
\title{get_meta_dt}
\usage{
get_meta_dt(
  seurat_obj,
  combine_source = "seurat_clusters",
  to_combine = list(west_island = c(15, 6), west_projection = c(13, 4), lonely_island =
    c(11), north_archipelago = c(9, 8, 14, 12), main_continent = c(7, 0, 2, 16, 2, 1, 3,
    5, 10)),
  cluster_vars = c("seurat_clusters", "meta_cluster"),
  manual_factors = list(orig.ident = c("wt", "df4")),
  reduction = "umap"
)
}
\arguments{
\item{seurat_obj}{an object of class Seurat to extract meta data from}

\item{to_combine}{}

\item{cluster_vars}{cluster variables, will be made factors with levels ordered by decreasing size.}

\item{manual_factors}{named list, names specify variables and entries specify factor levels to apply.}

\item{reduction}{string to specify reduction, must be in seurat_obj@reductions}
}
\value{

}
\description{
retrieves metadata data.table from Seurat object with reduction coordinates
added.  optional arguments all meta clustres to be specified by combining
an existing cluster definition and factor level specification for arbitrary
variables.
}
