% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_my_files.R
\name{get_my_files}
\alias{get_my_files}
\title{returns files from a root directory stored in sub directories per sample}
\usage{
get_my_files(root, repKey = 2:4, poolKey = 2:3)
}
\arguments{
\item{root}{directory containing directories to search}

\item{repKey}{parts of reps directories to use for naming. [-_\\.] delimitted.}

\item{poolKey}{parts of pooled directories to use for naming. [-_\\.] delimitted.}
}
\value{

}
\description{
returns files from a root directory stored in sub directories per sample
}
