% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scRNAseq.R
\name{get_rna_dt}
\alias{get_rna_dt}
\title{get_rna_dt}
\usage{
get_rna_dt(
  seurat_obj,
  sel_genes = NULL,
  raw_counts = FALSE,
  assay_name = "RNA"
)
}
\arguments{
\item{seurat_obj}{}

\item{sel_genes}{}

\item{raw_counts}{}

\item{assay_name}{}
}
\value{

}
\description{
retrieves RNA data from a Seurat object for selected genes in a tidy format.
}
