% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bigwig_correlation.R
\name{ggheatmap.dendro}
\alias{ggheatmap.dendro}
\title{Title}
\usage{
ggheatmap.dendro(mat, fill = "", title = "", subtitle = "")
}
\arguments{
\item{mat}{}

\item{fill}{}

\item{title}{}

\item{subtitle}{}
}
\value{

}
\description{
Title
}
