% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_load_STAR_RNA_counts.R
\name{guess_lib_from_file}
\alias{guess_lib_from_file}
\title{guess_lib_from_file}
\usage{
guess_lib_from_file(
  f,
  show_plots = FALSE,
  miss_cutoff = 1.2,
  hit_cutoff = 1.2,
  split_cutoff = 1.5
)
}
\arguments{
\item{f}{}

\item{show_plots}{}

\item{miss_cutoff}{}

\item{hit_cutoff}{}

\item{split_cutoff}{}
}
\value{

}
\description{
guess_lib_from_file
}
\examples{
files = dir("inst/extdata", pattern = "^test_[0-9].ReadsPerGene.out.tab", full = TRUE)
#unstranded
guess_lib_from_file(files[1], show_plots = TRUE)
guess_lib_from_file(files[2], show_plots = TRUE)
guess_lib_from_file(files[3], show_plots = TRUE)

#second strand
file.second_strand = dir("inst/extdata", pattern = "test_second.ReadsPerGene.out.tab", full = TRUE)
guess_lib_from_file(file.second_strand[1], show_plots = TRUE)
}
