% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_heatmap_with_sideplots.R
\name{h.cluster_data}
\alias{h.cluster_data}
\title{clust data.table accessor}
\usage{
h.cluster_data(ssvH2)
}
\arguments{
\item{ssvH2}{output from ssvHeatmap2()}
}
\value{
data supplied to geom_raster() with cluster assignments
}
\description{
clust data.table accessor
}
\examples{
h.cluster_data(ssvHeatmap2(heatmap_demo_matrix))
}
