% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_heatmap_with_sideplots.R
\name{h.cluster_members}
\alias{h.cluster_members}
\title{cluster members accessor}
\usage{
h.cluster_members(ssvH2)
}
\arguments{
\item{ssvH2}{output from ssvHeatmap2()}
}
\value{
list of cluster members
}
\description{
cluster members accessor
}
\examples{
h.cluster_members(ssvHeatmap2(heatmap_demo_matrix))
}
