% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_heatmap_with_sideplots.R
\name{h.final_plot}
\alias{h.final_plot}
\title{final_plot accessor}
\usage{
h.final_plot(ssvH2)
}
\arguments{
\item{ssvH2}{output from ssvHeatmap2()}
}
\value{
final assembly for heatmap figure, view with print()
}
\description{
final_plot accessor
}
\examples{
h.final_plot(ssvHeatmap2(heatmap_demo_matrix))
}
