% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_heatmap_with_sideplots.R
\name{h.plot_parts_grouped}
\alias{h.plot_parts_grouped}
\title{h.plot_parts_grouped}
\usage{
h.plot_parts_grouped(ssvH2)
}
\arguments{
\item{ssvH2}{output from ssvHeatmap2()}
}
\value{
intermediate assembly of heatmap components
}
\description{
h.plot_parts_grouped
}
\examples{
h.plot_parts_grouped(ssvHeatmap2(heatmap_demo_matrix))
}
