% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_heatmap_with_sideplots.R
\name{h.plot_parts_individual}
\alias{h.plot_parts_individual}
\title{h.plot_parts_individual}
\usage{
h.plot_parts_individual(ssvH2)
}
\arguments{
\item{ssvH2}{output from ssvHeatmap2()}
}
\value{
individual ggplot2 components of heatmap figure.
}
\description{
h.plot_parts_individual
}
\examples{
h.plot_parts_individual(ssvHeatmap2(heatmap_demo_matrix))
}
