% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_RNAseq_salmon.R
\name{is_url}
\alias{is_url}
\title{is_url}
\usage{
is_url(f)
}
\arguments{
\item{f}{a character possibly containing a URL}
}
\value{
TRUE if f looks like a URL
}
\description{
check if character f is a valid URL
}
\examples{
is_url("not_a_url")
is_url("ftp://is_a_url")
}
