% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_load_STAR_RNA_counts.R
\name{load_matrix_from_ReadsPerGene.out.tab}
\alias{load_matrix_from_ReadsPerGene.out.tab}
\title{load_matrix_from_ReadsPerGene.out.tab}
\usage{
load_matrix_from_ReadsPerGene.out.tab(files, lib_type = "guess")
}
\arguments{
\item{files}{}

\item{lib_type}{}
}
\value{

}
\description{
load_matrix_from_ReadsPerGene.out.tab
}
\examples{
files = dir("inst/extdata", pattern = "^test_[0-9].ReadsPerGene.out.tab", full = TRUE)
load_matrix_from_ReadsPerGene.out.tab(files)
}
