% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scRNA_fetch.R
\name{make_cell_dt_from_list}
\alias{make_cell_dt_from_list}
\title{Title}
\usage{
make_cell_dt_from_list(
  cell_assignment_list,
  cell_cluster_id_ = "seurat_clusters"
)
}
\arguments{
\item{cell_assignment_list}{list of character vector cell ids}

\item{cell_cluster_id_}{name to give cell cluster variable in data.table.  Default is "seurat_clusters".}
}
\value{
data.table equivalent of input lists.
}
\description{
Title
}
\examples{
cells_list = list(clust1 = c("a", "b"), clust2 = c("c"))
cells_dt = make_cell_dt_from_list(cells_list)
cells_dt
}
