% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{make_msigdb_TERM2GENE}
\alias{make_msigdb_TERM2GENE}
\title{make_msigdb_TERM2GENE}
\usage{
make_msigdb_TERM2GENE(
  species = "Mus musculus",
  category = "C8",
  subcategory = NULL
)
}
\arguments{
\item{species}{}

\item{category}{}

\item{subcategory}{}
}
\value{

}
\description{
Consult msigdbr for valid inputs
msigdbr::msigdbr_species()
msigdbr::msigdbr_collections()
}
\examples{
# Consult msigdbr for valid inputs
# msigdbr::msigdbr_species()
# msigdbr::msigdbr_collections()

make_msigdb_TERM2GENE("Mus musculus", "C8", NULL)
make_msigdb_TERM2GENE("Mus musculus", "C2", "CP:KEGG")
}
