% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{my_annotate}
\alias{my_annotate}
\title{Title}
\usage{
my_annotate(gr,
  ref_gr = rtracklayer::import.gff("~/gencode.v28.annotation.gtf.gz",
  feature.type = "transcript", format = "gtf"), gr_size = 2000)
}
\arguments{
\item{gr_size}{}
}
\description{
Title
}
