% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{my_clusterProfiler_fromGenes}
\alias{my_clusterProfiler_fromGenes}
\title{my_clusterProfiler_fromGenes}
\usage{
my_clusterProfiler_fromGenes(
  gene_lists,
  bg_genes = NULL,
  force_overwrite = FALSE,
  bfc = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{force_overwrite}{}
}
\value{

}
\description{
my_clusterProfiler_fromGenes
}
