% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{my_clusterProfiler_GO}
\alias{my_clusterProfiler_GO}
\title{my_clusterProfiler_fromGenes}
\usage{
my_clusterProfiler_GO(
  gene_lists,
  bg_genes = NULL,
  OrgDb = org.Hs.eg.db::org.Hs.eg.db,
  force_overwrite = FALSE,
  bfc = BiocFileCache::BiocFileCache(),
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.1
)
}
\arguments{
\item{force_overwrite}{}
}
\value{

}
\description{
my_clusterProfiler_fromGenes
}
