% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{my_clusterProfiler_kegg}
\alias{my_clusterProfiler_kegg}
\title{my_clusterProfiler_kegg}
\usage{
my_clusterProfiler_kegg(
  qgr,
  clust_assign,
  bg_genes = NULL,
  bfc = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{bg_genes}{}
}
\value{

}
\description{
my_clusterProfiler_kegg
}
