% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{my_clusterProfiler_MSigDB}
\alias{my_clusterProfiler_MSigDB}
\title{my_clusterProfiler_MSigDB}
\usage{
my_clusterProfiler_MSigDB(
  gene_lists,
  bg_genes = NULL,
  force_overwrite = FALSE,
  bfc = BiocFileCache::BiocFileCache(),
  organism = c("hsa", "mmu")[1],
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.1,
  msigdb_species = c("Homo sapiens", "Mus musculus")[1],
  msigdb_category = "C8",
  msigdb_subcategory = NULL,
  msigdb_TERM2GENE = NULL
)
}
\arguments{
\item{gene_lists}{}

\item{bg_genes}{}

\item{force_overwrite}{}

\item{bfc}{}

\item{organism}{}

\item{pvalueCutoff}{}

\item{qvalueCutoff}{}

\item{msigdb_species}{}

\item{msigdb_category}{}
}
\value{

}
\description{
my_clusterProfiler_MSigDB
}
