% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stranded_bamFetch.R
\name{myFetchStrandedBam}
\alias{myFetchStrandedBam}
\title{myFetchStrandedBam}
\usage{
myFetchStrandedBam(
  bam_files,
  qgr,
  win_size = 5,
  fragLens = NA,
  return_data.table = TRUE,
  flipStrand = FALSE,
  max_dupes = Inf,
  splice_strategy = c("none", "ignore", "add", "only")[1],
  anchor = "center"
)
}
\arguments{
\item{flipStrand}{}
}
\value{
data.table with strand specific reads
}
\description{
myFetchStrandedBam
}
