% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_stranded_bamFetch.R
\name{myPlotDiffClusters}
\alias{myPlotDiffClusters}
\title{different plots for showing strand differences}
\usage{
myPlotDiffClusters(diff_dt, clust_dt, cap = 30, fcap = 5)
}
\arguments{
\item{fcap}{}
}
\value{
list of ggplots
}
\description{
different plots for showing strand differences
}
