% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stranded_bamFetch.R
\name{myStrandDiff}
\alias{myStrandDiff}
\title{subtracts minus from plus strand}
\usage{
myStrandDiff(stranded_dt)
}
\arguments{
\item{cap}{}
}
\value{
data table with strand difference calculation
}
\description{
subtracts minus from plus strand
}
