% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_my_files.R
\name{parse_xls}
\alias{parse_xls}
\title{parse lines in xls}
\usage{
parse_xls(xls_file, key)
}
\arguments{
\item{xls_file}{xls file}

\item{key}{key to match}
}
\value{

}
\description{
parse lines in xls
}
