% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_my_files.R
\name{parse_xls_reads}
\alias{parse_xls_reads}
\title{parse lines in xls for treatment and control}
\usage{
parse_xls_reads(xls_file)
}
\arguments{
\item{xls_file}{xls file}
}
\description{
parse lines in xls for treatment and control
}
