% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alignment_mismatch.R
\name{plot_bam_mismatch_dt}
\alias{plot_bam_mismatch_dt}
\title{Title}
\usage{
plot_bam_mismatch_dt(all_bases_dt, mismatch_bases_dt, qgr, as_fraction = FALSE)
}
\arguments{
\item{all_bases_dt}{}

\item{mismatch_bases_dt}{}

\item{qgr}{}

\item{as_fraction}{}
}
\value{

}
\description{
Title
}
