% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scatter_with_side_density.R,
%   R/functions_scatterplot_with_side_density.R
\name{plot_scatter_side_density.assemble}
\alias{plot_scatter_side_density.assemble}
\title{synchronizes margins of central scatterplot to align density plots along x and y axis
places legend from primary scatter at top-right}
\usage{
plot_scatter_side_density.assemble(
  components,
  main_title = "",
  main_title.x = 0.02,
  main_title.y = 0.5,
  main_title.hjust = 0,
  main_title.vjust = 0.5,
  rel_widths = c(2, 1),
  rel_heights = c(1, 2)
)

plot_scatter_side_density.assemble(
  components,
  main_title = "",
  main_title.x = 0.02,
  main_title.y = 0.5,
  main_title.hjust = 0,
  main_title.vjust = 0.5,
  rel_widths = c(2, 1),
  rel_heights = c(1, 2)
)
}
\arguments{
\item{components}{}

\item{main_title}{}

\item{main_title.x}{}

\item{main_title.y}{}

\item{main_title.hjust}{}

\item{main_title.vjust}{}
}
\value{
a grob from cowplot::plot_grid


}
\description{
synchronizes margins of central scatterplot to align density plots along x and y axis
places legend from primary scatter at top-right

Title
}
