% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corrDiff.R
\name{plotHomerKnowResults}
\alias{plotHomerKnowResults}
\title{Plots knowresults from homer results directory}
\usage{
plotHomerKnowResults(
  homer_dir = "homer_strand_corr_results",
  dir_regex = "enriched.+Sorted",
  target_consensus = "AAACCACARM"
)
}
\arguments{
\item{homer_dir}{root directory of homer results}

\item{dir_regex}{regex to match sub-directories}

\item{target_consensus}{consensus to match from homer. default is AAACCACARM for RUNX1.}
}
\value{

}
\description{
Plots knowresults from homer results directory
}
