% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_fix_xlsx_dates.R
\name{read.table.fix_xlsx_dates}
\alias{read.table.fix_xlsx_dates}
\title{read.table.fix_xlsx_dates}
\usage{
read.table.fix_xlsx_dates(f, sep = ",", header = FALSE)
}
\arguments{
\item{f}{file path or data.frame}

\item{sep}{delimeter character for read.table, default is ,}

\item{header}{header logical for read.table, default is FALSE}
}
\value{
data.frame with dates converted to gene names
}
\description{
attempt to load file f and repair any gene names that excel has converted to dates
}
