% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_RNAseq_counts.R
\name{rnaseq_asses_strandedness}
\alias{rnaseq_asses_strandedness}
\title{rnaseq_asses_strandedness}
\usage{
rnaseq_asses_strandedness(bam_files, gtf_path, max_bams = 3)
}
\arguments{
\item{bam_files}{path to bam files}

\item{gtf_path}{path to gtf with exon info}

\item{max_bams}{only this many bams are used}
}
\value{
a grob of plots
}
\description{
rnaseq_asses_strandedness
}
