% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_RNAseq_salmon.R
\name{salmon_index_transcriptome}
\alias{salmon_index_transcriptome}
\title{salmon_index_transcriptome}
\usage{
salmon_index_transcriptome(
  seq_path = HG38_SEQ_PATH,
  gtf_path = HG38_v28_GTF_URL,
  gtf_to_fasta_path = GTF_TO_FASTA_PATH,
  salmon_path = SALMON_PATH,
  output_transcriptome = NULL,
  output_index = NULL,
  cache_path = "~/.cache",
  use_qsub = TRUE,
  do_submit = TRUE
)
}
\arguments{
\item{seq_path}{path to local file or url to assembly fasta (only tested with hg38).}

\item{gtf_path}{path to local file or url to gtf reference (only tested with gencode hg38 v28 gtf).}

\item{gtf_to_fasta_path}{local path to tophat2 gtf_to_fasta utility.}

\item{salmon_path}{path to main salmon executable.}

\item{output_transcriptome}{file path for transcriptome output.}

\item{output_index}{file path for index for transcriptome.}

\item{cache_path}{path to cache directory, will be created if necessary.}

\item{use_qsub}{whether to qsub a script. if FALSE will use a direct system command}

\item{do_submit}{if FALSE, will not perform qsub.}
}
\value{
list containing path to index and job_id if relevant
}
\description{
creates the salmon index for transcriptome created from specified assembly
fasta and gene reference gtf.
}
\examples{
# as this may download several gigs of data, do not run
\dontrun{
salmon_index_transcriptome()
}
}
