% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_RNAseq_salmon.R
\name{salmon_tx_quant}
\alias{salmon_tx_quant}
\title{Title}
\usage{
salmon_tx_quant(sf_files, gtf_path = HG38_v28_GTF_URL, cache_path = "~/.cache")
}
\arguments{
\item{cache_path}{}
}
\value{

}
\description{
Title
}
