% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scRNA_fetch.R
\name{ssvFetchBam.scRNA}
\alias{ssvFetchBam.scRNA}
\title{Title}
\usage{
ssvFetchBam.scRNA(
  file_path,
  qgr,
  cell_cluster_assignments,
  cell_cluster_id_ = "seurat_clusters",
  id_prefixes = "",
  unique_names = NULL,
  win_size = 50,
  win_method = c("sample", "summary")[1],
  summary_FUN = stats::weighted.mean,
  target_strand = c("*", "+", "-", "both")[1],
  flip_strand = FALSE,
  anchor = c("left", "left_unstranded", "center", "center_unstranded")[3],
  names_variable = "sample",
  return_data.table = FALSE,
  max_dupes = Inf,
  splice_strategy = c("none", "ignore", "add", "only", "splice_count")[2],
  n_cores = getOption("mc.cores", 1),
  return_unprocessed = FALSE,
  force_skip_centerFix = FALSE,
  force_barcodes = FALSE,
  min_seq_qual = 0,
  ...
)
}
\arguments{
\item{file_path}{character vector of file_path to load from. Alternatively,
file_path can be a data.frame or data.table whose first column is a
character vector of paths and additial columns will be used as metadata.}

\item{qgr}{Set of GRanges to query.  For valid results the width of each
interval should be identical and evenly divisible by \code{win_size}.}

\item{cell_cluster_assignments}{either a list of character vectors containing cell ids or a data.table.  If data.table one column bust be 'id' and contain cell barcodes, the other must match cell_cluster_id_ and contain cluster names.}

\item{cell_cluster_id_}{Attribute name of cell_cluster_assignments where cell barcodes are stored.}

\item{id_prefixes}{Character prefix to prepend to cell barcodes retrieved
from bam files such that they match id items in cell_cluster_assignments.
Either a single item to reuse or character vector of length equal to number
of bam files.  Default of "" adds nothing.}

\item{unique_names}{not used.  see cell_cluster_assignments for analogous functionality.}

\item{...}{}
}
\value{

}
\description{
Title
}
