% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_heatmap_with_sideplots.R
\docType{class}
\name{ssvH2-class}
\alias{ssvH2-class}
\alias{ssvH2}
\title{ssvH2}
\usage{
\S4method{initialize}{ssvH2}(.Object, final_plot, cluster_members,
  cluster_data, plot_parts_grouped, plot_parts_individual)
}
\description{
Stores results from ssvHeatmap2
}
