% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_alignment_mismatch.R
\name{ssvR_plot_bam_mismatch}
\alias{ssvR_plot_bam_mismatch}
\title{ssvR_plot_bam_mismatch}
\usage{
ssvR_plot_bam_mismatch(bamfile, qgr,
  gen_ref = BSgenome.Hsapiens.UCSC.hg38::Hsapiens, min_count = 5,
  as_fraction = FALSE)
}
\arguments{
\item{qgr}{}
}
\description{
ssvR_plot_bam_mismatch
}
\examples{
library(GenomicRanges)
qgr = GRanges("chr21:34815000-34835000")
bam_file = file.path("/slipstream/galaxy/uploads/working/qc_framework/output_drugs_with_merged_inputs/MCF10A_ctrl_input_pooled/MCF10A_ctrl_input_pooled.bam")
ssvR_plot_bam_mismatch(bam_file, qgr)
}
