% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot_ideogram.R
\name{ssvR_plot_ideogogram}
\alias{ssvR_plot_ideogogram}
\title{ssvR_plot_ideogogram}
\usage{
ssvR_plot_ideogogram(
  gen = "hg38",
  chr_to_show = paste0("chr", c(1:22, "X", "Y")),
  gr_highlights = NULL,
  bfc = BiocFileCache::BiocFileCache(),
  grIdeo = NULL,
  ideo_ymin = -1,
  ideo_ymax = 0,
  highlight_fill = "green",
  highlight_color = NA,
  highlight_alpha = 0.5,
  highlight_name = "highlight",
  facet_cols = 1,
  facet_by_row = FALSE,
  print_plot = TRUE,
  plot_title = ""
)
}
\arguments{
\item{highlight_alpha}{}
}
\value{

}
\description{
ssvR_plot_ideogogram
}
\examples{
ideo_res = ssvR_plot_ideogogram(facet_cols = 2)
}
