% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot_ideogram.R
\name{ssvR_plot_ideogogram_data}
\alias{ssvR_plot_ideogogram_data}
\title{ssvR_plot_ideogogram_data}
\usage{
ssvR_plot_ideogogram_data(
  data_dt,
  gen = "hg38",
  chr_to_show = paste0("chr", c(1:22, "X", "Y")),
  gr_highlights = NULL,
  bfc = BiocFileCache::BiocFileCache(),
  ideo_ymin = -1,
  ideo_ymax = 0,
  highlight_fill = "green",
  highlight_color = NA,
  highlight_alpha = 0.5,
  facet_cols = 1,
  facet_by_row = FALSE,
  data_ymin = 0,
  data_ymax = 1,
  print_plot = TRUE
)
}
\arguments{
\item{data_ymax}{}
}
\value{

}
\description{
ssvR_plot_ideogogram_data
}
\examples{
library(data.table)
f = system.file("extdata/pcaOut.PC1.chr2x.txt", package = "ssvRecipes")
dat = fread(f)
dat = dat[, .(x = (start+end)/2, y = PC1, seqnames = chr)]
ssvR_plot_ideogogram_data(dat, facet_cols = 2, chr_to_show = c("chr20", "chr21", "chr22"))
ssvR_plot_ideogogram_data(dat, facet_cols = 1, data_ymax = 10, chr_to_show = c("chr20", "chr21", "chr22"))
}
