% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{symbol2uniprot}
\alias{symbol2uniprot}
\title{symbol2uniprot}
\usage{
symbol2uniprot(x, OrgDb = org.Hs.eg.db::org.Hs.eg.db)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
symbol2uniprot
}
