% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_syncDims.R
\name{sync_height}
\alias{sync_height}
\title{sync ggplot heights}
\usage{
sync_height(my_plots)
}
\arguments{
\item{my_plots}{a list of ggplots}
}
\value{
a list of grobs wtih x margins all equal
}
\description{
sync ggplot heights
}
