% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_syncDims.R
\name{sync_width}
\alias{sync_width}
\title{sync ggplot widths}
\usage{
sync_width(my_plots)
}
\arguments{
\item{my_plots}{a list of ggplots}
}
\value{
a list of grobs wtih x margins all equal
}
\description{
sync ggplot widths
}
