% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tracks.R
\name{track_assembly}
\alias{track_assembly}
\title{Title}
\usage{
track_assembly(p_list, qgr, rel_heights = rep(1, length(p_list)))
}
\arguments{
\item{rel_heights}{}
}
\value{

}
\description{
Title
}
