% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_tracks.R
\name{track_ref}
\alias{track_ref}
\title{Title}
\usage{
track_ref(ref = "~/gencode.v28.annotation.gtf.gz", qgr, flip_x = FALSE,
  exon_height = 0.5, debug = FALSE)
}
\arguments{
\item{debug}{}
}
\description{
Title
}
