% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tracks.R
\name{track_rna}
\alias{track_rna}
\title{track_rna}
\usage{
track_rna(
  bams,
  qgr,
  fetch_fun = ssvRecipes::myFetchStrandedBam,
  flip_x = FALSE,
  flip_strand = FALSE,
  max_dupes = Inf,
  win_size = 100,
  strand_upsidedown = TRUE,
  ...
)
}
\arguments{
\item{strand_upsidedown}{}
}
\value{

}
\description{
track_rna
}
