% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_truncate_peak_file.R
\name{truncate_peak_file}
\alias{truncate_peak_file}
\title{truncate_peak_file}
\usage{
truncate_peak_file(
  np_file,
  order_VAR,
  out_dir = paste0(dirname(np_file), ".diffbind_input"),
  top_fraction = 0.1,
  top_n = NULL
)
}
\arguments{
\item{np_file}{path to .narrowPeak file or any .bed file using the same format}

\item{order_VAR}{variable to order by, typically signalValue or pValue}

\item{out_dir}{output location of truncated peaks, default is input file directory with ".diffbind_input" appended.}

\item{top_fraction}{fraction of top ranking peaks to retain. Default is .1}

\item{top_n}{number of top ranking peaks to retain. supersedes top_fraction if set. Default is NULL.}
}
\value{
path to truncated file
}
\description{
truncate_peak_file
}
\examples{
peaks2 = dir("/slipstream/galaxy/uploads/working/qc_framework/output/old_peaks/", 
             pattern = "H3K4ME3_pooled_peaks.narrowPeak$", full.names = TRUE)
peaks2.truncated_20percent = sapply(peaks2, truncate_peak_file, order_VAR= "signalValue", top_fraction = .2)
peaks2.truncated_5k = sapply(peaks2, truncate_peak_file, order_VAR= "pValue", top_n = 5e3)
}
