% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_hclust.R
\name{write_geneList_matrix}
\alias{write_geneList_matrix}
\title{write_geneList_matrix}
\usage{
write_geneList_matrix(gl, file)
}
\arguments{
\item{gl}{list of gene ids, must have cluster ids as naames.}

\item{file}{csv file to write to.}
}
\value{

}
\description{
write_geneList_matrix
}
