% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_file_helpers.R
\name{writeBigBed}
\alias{writeBigBed}
\title{Write gr to bigbed file}
\usage{
writeBigBed(gr, bbf, bedf = sub("\\\\.bb$", "\\\\.bed", bbf),
  chrSizes = "~/hg38_chrsizes.txt", cleanup_beds = TRUE)
}
\arguments{
\item{gr}{gr to save}

\item{bbf}{name of final output bigbed file}

\item{bedf}{name of intermediate bedfiles}

\item{chrSizes}{chrSizes file to use, default is hg38}

\item{cleanup_beds}{remove beds when done? TRUE}
}
\value{
path to big bed file
}
\description{
Write gr to bigbed file
}
