% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_multimapping.R
\name{writeFastaFromGR}
\alias{writeFastaFromGR}
\title{write a fasta file of specified read_size from reference sequence for alignment}
\usage{
writeFastaFromGR(gr, fasta_file, mc = parallel::detectCores() - 4,
  genome = BSgenome.Hsapiens.UCSC.hg38::Hsapiens, read_size = 75,
  wrap = 100)
}
\arguments{
\item{gr}{region to generate reads for}

\item{fasta_file}{path to output fasta file}

\item{mc}{number of cores to use. default is half of max.}

\item{genome}{BSgenome to use}

\item{read_size}{size of reads to make}
}
\description{
write a fasta file of specified read_size from reference sequence for alignment
}
