% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_multimapping.R
\name{writeFastaFromSeq}
\alias{writeFastaFromSeq}
\title{write a fasta file of specified read_size from reference sequence for alignment}
\usage{
writeFastaFromSeq(mySeq, fasta_file, mc = parallel::detectCores() - 4,
  read_size = 75, wrap = 100)
}
\arguments{
\item{fasta_file}{path to output fasta file}

\item{mc}{number of cores to use. default is half of max.}

\item{read_size}{size of reads to make}

\item{seq}{the seq}

\item{genome}{BSgenome to use}
}
\description{
write a fasta file of specified read_size from reference sequence for alignment
}
