% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{annotate_gr_clusters}
\alias{annotate_gr_clusters}
\title{annotate_gr_clusters}
\usage{
annotate_gr_clusters(qgr, clust_assign, tss_gr, max_dist = 5000)
}
\arguments{
\item{qgr}{}

\item{clust_assign}{A data.frame derived from ssvSignalClustering output (clust_assign = unique(clust_dt[, .(id, cluster_id)]))}
}
\value{

}
\description{
annotate_gr_clusters
}
