% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{my_clusterProfiler_KEGG}
\alias{my_clusterProfiler_KEGG}
\title{my_clusterProfiler_KEGG}
\usage{
my_clusterProfiler_KEGG(
  gene_lists,
  bg_genes = NULL,
  force_overwrite = FALSE,
  bfc = BiocFileCache::BiocFileCache(),
  organism = c("hsa", "mmu")[1],
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.1
)
}
\arguments{
\item{force_overwrite}{}
}
\value{

}
\description{
my_clusterProfiler_KEGG
}
