% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{my_clusterProfiler_KEGG}
\alias{my_clusterProfiler_KEGG}
\title{my_clusterProfiler_KEGG}
\usage{
my_clusterProfiler_KEGG(
  gene_lists,
  bg_genes = NULL,
  gene_lists.as_is = FALSE,
  bg_genes.as_is = FALSE,
  OrgDb = org.Hs.eg.db::org.Hs.eg.db,
  force_overwrite = FALSE,
  bfc = BiocFileCache::BiocFileCache(),
  organism = c("hsa", "mmu")[1],
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.1
)
}
\arguments{
\item{gene_lists}{KEGG requires UNIPROT, gene_lists is expected to be SYMBOL.  If already UNIPROT set gene_lists.as_is = TRUE.}

\item{bg_genes}{KEGG requires UNIPROT, bg_genes is expected to be SYMBOL.  If already UNIPROT set bg_genes.as_is = TRUE.}

\item{force_overwrite}{}
}
\value{

}
\description{
my_clusterProfiler_KEGG
}
