% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clusterProfiler.R
\name{my_clusterProfiler_kegg}
\alias{my_clusterProfiler_kegg}
\title{Title}
\usage{
my_clusterProfiler_kegg(qgr, clust_assign, bg_genes = NULL)
}
\arguments{
\item{bg_genes}{}
}
\description{
Title
}
