% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_heatmap_with_sideplots.R
\docType{methods}
\name{plot,ssvH2,ANY-method}
\alias{plot,ssvH2,ANY-method}
\alias{plot,ssvH2-method}
\title{plot method for ssvH2}
\usage{
\S4method{plot}{ssvH2,ANY}(x)
}
\description{
plot method for ssvH2
}
