% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_scatter_with_side_density.R
\name{plot_scatter_side_density.assemble}
\alias{plot_scatter_side_density.assemble}
\title{synchronizes margins of central scatterplot to align density plots along x and y axis
places legend from primary scatter at top-right}
\usage{
plot_scatter_side_density.assemble(components, main_title = "",
  main_title.x = 0.02, main_title.y = 0.5, main_title.hjust = 0,
  main_title.vjust = 0.5)
}
\arguments{
\item{components}{list with ggplot items: scatter, x_density, and y_density}

\item{main_title}{optional title to add at top}

\item{main_title.x}{x coord of title, 0-1, default is 0.02}

\item{main_title.y}{y coord of title, 0-1, default is 0.5}

\item{main_title.hjust}{hjust of title, default is 0}

\item{main_title.vjust}{vjust of title, default is 0.5}
}
\value{
a grob from cowplot::plot_grid
}
\description{
synchronizes margins of central scatterplot to align density plots along x and y axis
places legend from primary scatter at top-right
}
