% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_heatmap_with_sideplots.R
\name{ssvHeatmap2}
\alias{ssvHeatmap2}
\title{seqsetvis improved Heatmap}
\usage{
ssvHeatmap2(mat, treatment_ordering = NULL, replicate_ordering = NULL,
  treatment_space_size = 0.1, treatment_label_size = 8,
  main_heights = c(8, 1, 2, 2), main_widths = c(2, 0.1, 0.2, 0.2, 1.8),
  cluster_ = "cluster_id", treatment_ = "sample", replicate_ = "x",
  column_ = "column", row_ = "id", fill_ = "y", fill_label = fill_,
  nclust = 5, main_title = "Heatmap", main_title.x = 0.02,
  main_title.y = 0.98, main_title.hjust = 0, main_title.vjust = 1,
  side_plot_type = c("lines1", "lines2", "bars1", "bars2")[1],
  side_plot_FUN = NULL, side_plot_ylab = "",
  side_plot_ylab_width = 0.1,
  heatmap_colors = rev(seqsetvis::safeBrew(5, pal = "spectral")),
  symm_colors = TRUE, side_plot_colors = NULL)
}
\arguments{
\item{mat}{numeric matrix with row and column names set.  see details.}

\item{treatment_ordering}{order used for columns and side plots.  default NULL
causes order of supplied data to be used.}

\item{replicate_ordering}{order used for sub columns and side plots ticks.
default NULL causes order of supplied data to be used.}

\item{treatment_space_size}{numeric >=0. size of white space separating heatmap
groups. Default is 0.1}

\item{treatment_label_size}{numeric > 0. size of text to label groups with, default is 8.}

\item{main_heights}{numeric of length 4.  controls size of heatmap
body, axis ticks, group labels, and scale portion.}

\item{main_widths}{numeric of length 5.  controls size of heatmap
body, spacer, cluster bars, cluster connector, and aggregate plots}

\item{treatment_}{name of conditions/treatment groups.  appears in plots.}

\item{replicate_}{name of replicate identifier.  appears in plots.}

\item{column_}{name of heatmap column variable - no effect if supplying mat
as matrix.}

\item{row_}{name of heatmap row variable - no effect if supplying mat as
matrix.}

\item{fill_}{name of heatmap fill variable - no effect if supplying mat as
matrix.}

\item{nclust}{numeric number of clusters}

\item{main_title}{title appearing above assembled figure.}

\item{side_plot_type}{plotting method for aggregate side plots.
one of c("lines1", "lines2", "bars1", "bars2").}

\item{side_plot_FUN}{overrides sideplot type. supplied function is executed
once per cluster and must return a ggplot. call is:
side_plot_FUN(clust, i, fill_, replicate_, treatment_, cluster_, side_plot_colors)
Where clust is a tidy data.table containing 4 variables_, i is cluster id.}

\item{heatmap_colors}{colors for heatmap, default uses spectral palette
from color brewer.}

\item{side_plot_colors}{colors for side_plot.
must be same length as number of treatment groups. default uses dark2
palette from  color brewer.}
}
\value{
a ssvH2 object.  print() this object for help.
}
\description{
An upgrade to setsetvis' ssvSignalHeatmap() function
accepts a standard wide matrix with some colname an rowname assumptions.
}
\details{
colnames must contain a single "_".

Condition or treatment groups should
precede the "_".

A unique replicate identifier should follow the "_".
Replicate identifiers should be unique within each treatment group.

valid colnames:

A_1, A_2, B_1, B_2

A_1, A_2, A_3

A_1, B_1, C_1, C_2

rownames must be set and be unique.
}
\examples{
res = ssvHeatmap2(heatmap_demo_matrix)
print(res)
plot(res)

res2 = ssvHeatmap2(heatmap_demo_matrix, side_plot_ylab = 'y-value')
plot(res2)
}
