% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_RNAseq_counts.R
\name{star_align_fastq_PE}
\alias{star_align_fastq_PE}
\title{Title}
\usage{
star_align_fastq_PE(r1_fastq_paths, r2_fastq_paths = NULL,
  pair_key = "P", r1_to_r2_FUN = function(r1) {    
  sub(paste0(pair_key, "1.fastq"), paste0(pair_key, "2.fastq"), r1) },
  index_path = HG38_STAR_INDEX, gtf_path = HG38_v28_GTF_URL,
  star_path = STAR_PATH, cache_path = "~/.cache", n_cores = 8,
  hold_jids = NA, out_path = file.path(getwd(), "alignment"),
  output_prefix = paste0(sub(paste0("_", pair_key, "1.fastq"), "",
  basename(r1_fastq_paths)), ".STAR."), do_submit = TRUE)
}
\arguments{
\item{r1_fastq_paths}{paths to r1 files, mandatory}

\item{r2_fastq_paths}{paths to r2 files, will try to derive by calling r1_to_r2_FUN(r1_fastq_paths)}

\item{r1_to_r2_FUN}{optional function if r2 files no supplied}

\item{gtf_path}{path (local or url) to gtf}

\item{star_path}{path to star executable}

\item{cache_path}{cache location}

\item{n_cores}{number of threads to use}

\item{hold_jids}{job ids to hold for, default NA is none.}

\item{out_path}{directory to output to}

\item{output_prefix}{custom output directory per fastq_paths. default is basename
of fastq_paths}

\item{do_submit}{if FALSE, qsub is skipped but submit scripts remain.
Default is TRUE.}
}
\value{
list length 2 of output dirs and job ids
}
\description{
Title
}
