% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tracks.R
\name{track_chip}
\alias{track_chip}
\title{Title}
\usage{
track_chip(
  bw_files,
  qgr,
  flip_x = FALSE,
  nwin = 100,
  win_FUN = c("mean", "max")[1],
  nspline = 10,
  debug = FALSE,
  show_lines = TRUE,
  show_fill = FALSE
)
}
\arguments{
\item{show_fill}{}
}
\value{

}
\description{
Title
}
