% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tracks.R
\name{track_ref}
\alias{track_ref}
\title{track_ref}
\usage{
track_ref(
  ref = "~/gencode.v28.annotation.gtf.gz",
  qgr,
  flip_x = FALSE,
  exon_height = 0.5,
  intron_thickness = 2,
  show_tss = FALSE,
  tss_size = 3,
  tss_color = "red",
  return_data = FALSE
)
}
\arguments{
\item{debug}{}
}
\value{

}
\description{
track_ref
}
\examples{
ref = rtracklayer::import.gff(ref, format =  "gtf", feature.type = "exon")
qgr = range(subset(ref, gene_name == "GREB1"))
qgr
track_ref(ref, qgr)

track_ref(ref, qgr, show_tss = TRUE)

}
